refactor: split compose build dedupe
